Cast for clang-cl 64 bit build warnings in unittests
R=kjellander@chromium.org
BUG=libyuv:649

Review URL: https://codereview.chromium.org/2414763002 .

+++ unit_test/video_common_test.cc
@@ -42,23 +42,23 @@ static bool TestValidFourCC(uint32 fourcc, int bpp) {
 }
 
 TEST_F(LibYUVBaseTest, TestCanonicalFourCC) {
-  EXPECT_EQ(FOURCC_I420, CanonicalFourCC(FOURCC_IYUV));
-  EXPECT_EQ(FOURCC_I420, CanonicalFourCC(FOURCC_YU12));
-  EXPECT_EQ(FOURCC_I422, CanonicalFourCC(FOURCC_YU16));
-  EXPECT_EQ(FOURCC_I444, CanonicalFourCC(FOURCC_YU24));
-  EXPECT_EQ(FOURCC_YUY2, CanonicalFourCC(FOURCC_YUYV));
-  EXPECT_EQ(FOURCC_YUY2, CanonicalFourCC(FOURCC_YUVS));
-  EXPECT_EQ(FOURCC_UYVY, CanonicalFourCC(FOURCC_HDYC));
-  EXPECT_EQ(FOURCC_UYVY, CanonicalFourCC(FOURCC_2VUY));
-  EXPECT_EQ(FOURCC_MJPG, CanonicalFourCC(FOURCC_JPEG));
-  EXPECT_EQ(FOURCC_MJPG, CanonicalFourCC(FOURCC_DMB1));
-  EXPECT_EQ(FOURCC_RAW,  CanonicalFourCC(FOURCC_RGB3));
-  EXPECT_EQ(FOURCC_24BG, CanonicalFourCC(FOURCC_BGR3));
-  EXPECT_EQ(FOURCC_BGRA, CanonicalFourCC(FOURCC_CM32));
-  EXPECT_EQ(FOURCC_RAW,  CanonicalFourCC(FOURCC_CM24));
-  EXPECT_EQ(FOURCC_RGBO, CanonicalFourCC(FOURCC_L555));
-  EXPECT_EQ(FOURCC_RGBP, CanonicalFourCC(FOURCC_L565));
-  EXPECT_EQ(FOURCC_RGBO, CanonicalFourCC(FOURCC_5551));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_I420), CanonicalFourCC(FOURCC_IYUV));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_I420), CanonicalFourCC(FOURCC_YU12));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_I422), CanonicalFourCC(FOURCC_YU16));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_I444), CanonicalFourCC(FOURCC_YU24));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_YUY2), CanonicalFourCC(FOURCC_YUYV));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_YUY2), CanonicalFourCC(FOURCC_YUVS));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_UYVY), CanonicalFourCC(FOURCC_HDYC));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_UYVY), CanonicalFourCC(FOURCC_2VUY));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_MJPG), CanonicalFourCC(FOURCC_JPEG));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_MJPG), CanonicalFourCC(FOURCC_DMB1));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_RAW),  CanonicalFourCC(FOURCC_RGB3));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_24BG), CanonicalFourCC(FOURCC_BGR3));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_BGRA), CanonicalFourCC(FOURCC_CM32));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_RAW),  CanonicalFourCC(FOURCC_CM24));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_RGBO), CanonicalFourCC(FOURCC_L555));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_RGBP), CanonicalFourCC(FOURCC_L565));
+  EXPECT_EQ(static_cast<uint32>(FOURCC_RGBO), CanonicalFourCC(FOURCC_5551));
 }
 
 TEST_F(LibYUVBaseTest, TestFourCC) {
